+++ test/landsandboat_simulation.h
@@ -2,18 +2,18 @@
 
 #include <atomic>
 #include <chrono>
+#include <memory>
 #include <random>
 #include <vector>
-#include <memory>
 
 namespace LandSandBoatSimulation
 {
 
 auto getRandomFloat() -> float
 {
-    static std::random_device rd;
+    static std::random_device                    rd;
     static std::uniform_real_distribution<float> distribution(0.0f, 1.0f);
-    static std::mt19937 generator(rd());
+    static std::mt19937                          generator(rd());
     return distribution(generator);
 }
 
@@ -25,46 +25,56 @@ auto getRandomFloat() -> float
 struct position_t
 {
     float x, y, z;
-    
-    position_t() : x(0.0f), y(0.0f), z(0.0f) {}
-    position_t(float x_, float y_, float z_) : x(x_), y(y_), z(z_) {}
-    
+
+    position_t()
+    : x(0.0f)
+    , y(0.0f)
+    , z(0.0f)
+    {
+    }
+    position_t(float x_, float y_, float z_)
+    : x(x_)
+    , y(y_)
+    , z(z_)
+    {
+    }
+
     // Basic vector operations
     position_t operator+(const position_t& other) const
     {
         return position_t(x + other.x, y + other.y, z + other.z);
     }
-    
+
     position_t operator-(const position_t& other) const
     {
         return position_t(x - other.x, y - other.y, z - other.z);
     }
-    
+
     float distance(const position_t& other) const
     {
         float dx = x - other.x;
         float dy = y - other.y;
         float dz = z - other.z;
-        return std::sqrt(dx*dx + dy*dy + dz*dz);
+        return std::sqrt(dx * dx + dy * dy + dz * dz);
     }
 };
 
 // Entity data structure (simplified from LandSandBoat)
 struct Entity
 {
-    uint32_t id;
-    position_t position;
-    position_t target_position;
+    size_t                  id;
+    position_t              position;
+    position_t              target_position;
     std::vector<position_t> path;
-    bool is_moving;
-    bool needs_pathfinding;
-    
-    Entity(uint32_t entity_id) 
-        : id(entity_id)
-        , position(0.0f, 0.0f, 0.0f)
-        , target_position(0.0f, 0.0f, 0.0f)
-        , is_moving(false)
-        , needs_pathfinding(false)
+    bool                    is_moving;
+    bool                    needs_pathfinding;
+
+    Entity(size_t entity_id)
+    : id(entity_id)
+    , position(0.0f, 0.0f, 0.0f)
+    , target_position(0.0f, 0.0f, 0.0f)
+    , is_moving(false)
+    , needs_pathfinding(false)
     {
     }
 };
@@ -72,19 +82,22 @@ struct Entity
 // Zone data structure (simplified from LandSandBoat)
 struct Zone
 {
-    uint32_t zone_id;
+    size_t                               zone_id;
     std::vector<std::unique_ptr<Entity>> entities;
-    std::atomic<int> active_entities{0};
-    
-    Zone(uint32_t id) : zone_id(id) {}
-    
-    void addEntity(uint32_t entity_id)
+    std::atomic<int>                     active_entities{ 0 };
+
+    Zone(size_t id)
+    : zone_id(id)
+    {
+    }
+
+    void addEntity(size_t entity_id)
     {
         entities.emplace_back(std::make_unique<Entity>(entity_id));
         active_entities.fetch_add(1);
     }
-    
-    Entity* getEntity(uint32_t entity_id)
+
+    Entity* getEntity(size_t entity_id)
     {
         for (auto& entity : entities)
         {
@@ -104,26 +117,25 @@ inline auto co_simulatePathfinding(const position_t& start, const position_t& en
 {
     // Simulate actual pathfinding work (5ms as mentioned in the user's requirements)
     std::this_thread::sleep_for(std::chrono::milliseconds(5));
-    
+
     // Generate a realistic path with waypoints
     std::vector<position_t> path;
     path.push_back(start);
-    
+
     // Add some intermediate waypoints (simulating navmesh pathfinding)
-    float distance = start.distance(end);
-    int waypoints = static_cast<int>(distance / 10.0f); // Waypoint every 10 units
-    
+    float distance  = start.distance(end);
+    int   waypoints = static_cast<int>(distance / 10.0f); // Waypoint every 10 units
+
     for (int i = 1; i < waypoints && i < 5; ++i) // Max 5 waypoints
     {
-        float t = static_cast<float>(i) / static_cast<float>(waypoints);
+        float      t = static_cast<float>(i) / static_cast<float>(waypoints);
         position_t waypoint(
             start.x + (end.x - start.x) * t,
             start.y + (end.y - start.y) * t,
-            start.z + (end.z - start.z) * t
-        );
+            start.z + (end.z - start.z) * t);
         path.push_back(waypoint);
     }
-    
+
     path.push_back(end);
     co_return path;
 }
@@ -139,17 +151,16 @@ inline auto simulatePathfinding(const position_t& start, const position_t& end) 
     path.push_back(start);
 
     // Add some intermediate waypoints (simulating navmesh pathfinding)
-    float distance = start.distance(end);
-    int waypoints = static_cast<int>(distance / 10.0f); // Waypoint every 10 units
+    float distance  = start.distance(end);
+    int   waypoints = static_cast<int>(distance / 10.0f); // Waypoint every 10 units
 
     for (int i = 1; i < waypoints && i < 5; ++i) // Max 5 waypoints
     {
-        float t = static_cast<float>(i) / static_cast<float>(waypoints);
+        float      t = static_cast<float>(i) / static_cast<float>(waypoints);
         position_t waypoint(
             start.x + (end.x - start.x) * t,
             start.y + (end.y - start.y) * t,
-            start.z + (end.z - start.z) * t
-        );
+            start.z + (end.z - start.z) * t);
         path.push_back(waypoint);
     }
 
@@ -181,13 +192,13 @@ inline auto co_simulateEntityMovement(Entity* entity) -> AsyncTask<void>
 {
     // Simulate movement processing (2-5ms)
     std::this_thread::sleep_for(std::chrono::milliseconds(1 + (entity->id % 3)));
-    
+
     if (!entity->path.empty())
     {
         // Move towards next waypoint
         entity->position = entity->path[0];
         entity->path.erase(entity->path.begin());
-        
+
         if (entity->path.empty())
         {
             entity->is_moving = false;
@@ -268,22 +279,22 @@ inline auto co_processEntityAI(Entity* entity) -> AsyncTask<void>
     if (needs_movement && !entity->is_moving)
     {
         position_t target(getRandomFloat(), getRandomFloat(), getRandomFloat());
-        entity->target_position = target;
+        entity->target_position   = target;
         entity->needs_pathfinding = true;
-        
+
         // This is the key suspension point - pathfinding runs on worker thread
         entity->path = co_await co_pathfinding(entity->position, target);
-        
-        entity->is_moving = true;
+
+        entity->is_moving         = true;
         entity->needs_pathfinding = false;
     }
-    
+
     // Process movement
     if (entity->is_moving)
     {
         co_await co_simulateEntityMovement(entity);
     }
-    
+
     // Update database
     co_await co_simulateDatabaseUpdate(entity);
 
@@ -301,7 +312,7 @@ inline auto co_processZoneEntities(Zone* zone) -> AsyncTask<void>
             co_await co_processEntityAI(entity.get());
         }
     }
-    
+
     // Send network updates
     for (auto& entity : zone->entities)
     {
@@ -310,7 +321,7 @@ inline auto co_processZoneEntities(Zone* zone) -> AsyncTask<void>
             co_await co_simulateNetworkPacket(entity.get());
         }
     }
-    
+
     co_return;
 }
 
@@ -321,17 +332,17 @@ inline auto co_zoneServerTick(Zone* zone) -> Task<void>
 
     // Zone initialization work (like CZone::ZoneServer)
     std::this_thread::sleep_for(std::chrono::milliseconds(1));
-    
+
     // Process all entities in the zone
     co_await co_processZoneEntities(zone);
 
     // Zone cleanup and state updates
     std::this_thread::sleep_for(std::chrono::milliseconds(1));
 
-    const auto end = std::chrono::high_resolution_clock::now();
+    const auto end      = std::chrono::high_resolution_clock::now();
     const auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start);
     coro_log("Zone server tick duration: " + std::to_string(duration.count()) + "ms");
-    
+
     co_return;
 }
 
@@ -356,13 +367,13 @@ inline auto processEntityAI(Entity* entity) -> Task<void>
     if (needs_movement && !entity->is_moving)
     {
         position_t target(getRandomFloat(), getRandomFloat(), getRandomFloat());
-        entity->target_position = target;
+        entity->target_position   = target;
         entity->needs_pathfinding = true;
 
         // This is the key suspension point - pathfinding runs on worker thread
         entity->path = co_await pathfinding(entity->position, target);
 
-        entity->is_moving = true;
+        entity->is_moving         = true;
         entity->needs_pathfinding = false;
     }
 
@@ -416,7 +427,7 @@ inline auto zoneServerTick(Zone* zone) -> Task<void>
     // Zone cleanup and state updates
     std::this_thread::sleep_for(std::chrono::milliseconds(1));
 
-    const auto end = std::chrono::high_resolution_clock::now();
+    const auto end      = std::chrono::high_resolution_clock::now();
     const auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end - start);
     coro_log("Zone server tick duration: " + std::to_string(duration.count()) + "ms");
 
@@ -428,15 +439,15 @@ inline auto zoneServerTick(Zone* zone) -> Task<void>
 //
 
 // Create a realistic zone with entities
-inline auto createRealisticZone(uint32_t zone_id, uint32_t num_entities) -> std::unique_ptr<Zone>
+inline auto createRealisticZone(size_t zone_id, size_t num_entities) -> std::unique_ptr<Zone>
 {
     auto zone = std::make_unique<Zone>(zone_id);
-    
-    for (uint32_t i = 0; i < num_entities; ++i)
+
+    for (size_t i = 0U; i < num_entities; ++i)
     {
-        zone->addEntity(i + 1);
+        zone->addEntity(i + 1U);
     }
-    
+
     return zone;
 }
